Create gridWrapper in attempt to extract muuri specific stuff. Wont work since it's fairly tightly coupled :p
