+++ src/index.html
@@ -17,7 +17,7 @@
 
 <div class="top-bar">
     <div id="logo1" style="order: 1;"><a href="https://github.com/clanghout/ctab-page" target="_blank"
-                                         title="to github page">CTab</a>
+                                         title="to github page" rel="noopener noreferrer">CTab</a>
     </div>
     <div style="order: 2">
         <div class="sorting-control">
@@ -87,7 +87,7 @@
     <div class="settings-pane-main">
         <div class="setting">
             <label>
-                Weather API key (<a href="https://openweathermap.org/appid" target="_blank">openweathermap</a>)
+                Weather API key (<a href="https://openweathermap.org/appid" target="_blank" rel="noopener noreferrer">openweathermap</a>)
                 <input id="weather-API-key" name="weather-API-key" type="text">
             </label>
         </div>
